# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "blog-editor API",
    "version": "1",
    "description": "blog-editor 백오피스 API. 이 파일은 apps/editor/api/src/contract에서 내보낸 것이다 — 손으로 고치지 않고 `node apps/editor/api/src/contract/write-openapi.ts`로 다시 쓴다. zod의 refine 규칙(스티커 합계, 글꼴별 두께 등)은 JSON Schema로 표현되지 않으며, 어기면 400 SchemaErrorBody다. `/mcp`와 OAuth 경로는 MCP 프로토콜이라 이 계약 밖이다."
  },
  "paths": {
...
    "/api/posts": {
      "get": {
        "operationId": "listPosts",
        "tags": ["posts"],
        "summary": "글 목록",
        "description": "초안과 발행 글 모두, date 최신순.",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "글 요약 목록",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostList"
                }
              }
            }
          },
          "401": {
            "description": "로그인이 필요하다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/{slug}": {
      "get": {
        "operationId": "getPost",
        "tags": ["posts"],
        "summary": "글 읽기",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "글 주소 — 소문자 · 숫자 · 하이픈",
            "schema": {
              "type": "string",
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "저장 형식 그대로",
            "headers": {
              "ETag": {
                "description": "revision — 고쳐 저장할 때 If-Match로 돌려준다",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostFile"
                }
              }
            }
          },
          "400": {
            "description": "주소 모양이 틀렸다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageBody"
                }
              }
            }
          },
          "401": {
            "description": "로그인이 필요하다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageBody"
                }
              }
            }
          },
          "404": {
            "description": "글이 없다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageBody"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "savePost",
        "tags": ["posts"],
        "summary": "글 저장(새 글 · 고치기 · 발행)",
        "description": "새 글은 If-None-Match: *, 고치기는 If-Match: <ETag>. 발행은 meta.draft를 false로 저장하는 것이다(발행 전용 경로 없음). 본문의 doc은 저장 전에 정규형으로 바뀐다.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "글 주소 — 소문자 · 숫자 · 하이픈",
            "schema": {
              "type": "string",
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "description": "`*` — 새 글일 때만 저장. 그 주소에 글이 이미 있으면 409",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "description": "읽을 때 받은 ETag — 그 뒤 다른 곳에서 고쳤거나 그 주소에 글이 없으면 409",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "저장 형식",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostFile"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "고쳐 저장됨",
            "headers": {
              "ETag": {
                "description": "새 revision",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveResult"
                }
              }
            }
          },
          "201": {
            "description": "새로 저장됨",
            "headers": {
              "ETag": {
                "description": "새 revision",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveResult"
                }
              }
            }
          },
          "400": {
            "description": "주소 모양이 틀렸거나(MessageBody) 본문이 JSON이 아니거나 스키마에 맞지 않는다(SchemaErrorBody)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveBadRequestBody"
                }
              }
            }
          },
          "401": {
            "description": "로그인이 필요하다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageBody"
                }
              }
            }
          },
          "409": {
            "description": "revision이 맞지 않는다 — If-Match: 다른 곳에서 먼저 고쳤거나 글이 없다 · If-None-Match: *: 그 주소에 글이 이미 있다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageBody"
                }
              }
            }
          },
          "428": {
            "description": "If-None-Match · If-Match가 둘 다 없다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/public/posts": {
      "get": {
        "operationId": "listPublicPosts",
        "tags": ["public"],
        "summary": "공개 글 목록(사이트 빌드용)",
        "description": "발행 글만, 렌더된 HTML 포함. 초안이 섞이면 응답 대신 500이다(public-posts-api).",
        "security": [],
        "responses": {
          "200": {
            "description": "발행 글",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicPosts"
                }
              }
            }
          },
          "500": {
            "description": "응답이 공개 계약에 맞지 않는다(초안이 섞였다) — Hono 기본 오류 응답",
            "content": {
              "text/plain": {}
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-session"
      }
    },
    "schemas": {
      "PostFile": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "number",
            "const": 1
          },
          "meta": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string",
                "minLength": 1,
                "maxLength": 80
              },
              "description": {
                "type": "string",
                "minLength": 1,
                "maxLength": 160
              },
              "date": {
                "type": "string",
                "format": "date",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
              },
              "updated": {
                "type": "string",
                "format": "date",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
              },
              "category": {
                "type": "string",
                "description": "워크스페이스 설정의 카테고리 중 하나 — 목록은 설정이 주고, 서버의 zod가 최종 검증한다"
              },
              "draft": {
                "type": "boolean"
              },
              "image": {
                "type": "string",
                "pattern": "^\\/images\\/[a-z0-9-]+\\.(?:webp|png|jpg|jpeg|gif)$"
              },
              "source": {
                "anyOf": [
                  {
                    "type": "string",
                    "enum": ["editor", "claude", "chatgpt"]
                  },
                  {
                    "type": "string",
                    "pattern": "^token:[a-z0-9-]{1,32}$"
                  }
                ]
              }
            },
            "required": ["title", "description", "date", "category", "draft", "source"],
            "additionalProperties": false
          },
          "doc": {
            "$ref": "#/components/schemas/Doc"
          }
        },
        "required": ["schemaVersion", "meta", "doc"],
        "additionalProperties": false
      },
...
      "PostList": {
        "type": "object",
        "properties": {
          "posts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "slug": {
                  "type": "string",
                  "maxLength": 80,
                  "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
                },
                "title": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 80
                },
                "date": {
                  "type": "string",
                  "format": "date",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
                },
                "updated": {
                  "type": "string",
                  "format": "date",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
                },
                "category": {
                  "type": "string",
                  "description": "워크스페이스 설정의 카테고리 중 하나 — 목록은 설정이 주고, 서버의 zod가 최종 검증한다"
                },
                "draft": {
                  "type": "boolean"
                },
                "source": {
                  "anyOf": [
                    {
                      "type": "string",
                      "enum": ["editor", "claude", "chatgpt"]
                    },
                    {
                      "type": "string",
                      "pattern": "^token:[a-z0-9-]{1,32}$"
                    }
                  ]
                }
              },
              "required": ["slug", "title", "date", "category", "draft", "source"],
              "additionalProperties": false
            }
          }
        },
        "required": ["posts"],
        "additionalProperties": false
      },
      "PublicPosts": {
        "type": "object",
        "properties": {
          "postCssUrl": {
            "type": "string",
            "minLength": 1
          },
          "posts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "slug": {
                  "type": "string",
                  "maxLength": 80,
                  "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
                },
                "title": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 80
                },
                "description": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 160
                },
                "date": {
                  "type": "string",
                  "format": "date",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
                },
                "updated": {
                  "type": "string",
                  "format": "date",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
                },
                "category": {
                  "type": "string",
                  "description": "워크스페이스 설정의 카테고리 중 하나 — 목록은 설정이 주고, 서버의 zod가 최종 검증한다"
                },
                "draft": {
                  "type": "boolean",
                  "const": false
                },
                "image": {
                  "type": "string",
                  "format": "uri"
                },
                "html": {
                  "type": "string",
                  "format": "starts_with",
                  "pattern": "^<div class=\"post-body\">.*"
                }
              },
              "required": ["slug", "title", "description", "date", "category", "draft", "html"],
              "additionalProperties": false
            }
          }
        },
        "required": ["postCssUrl", "posts"],
        "additionalProperties": false
      },
...
      "SaveResult": {
        "type": "object",
        "properties": {
          "revision": {
            "type": "string"
          }
        },
        "required": ["revision"],
        "additionalProperties": false
      },
...
      "MessageBody": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": ["message"],
        "additionalProperties": false
      },
      "SchemaErrorBody": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "array",
                  "items": {
                    "type": ["string", "number"]
                  }
                },
                "message": {
                  "type": "string"
                }
              },
              "required": ["path", "message"],
              "additionalProperties": false
            }
          }
        },
        "required": ["message", "issues"],
        "additionalProperties": false
      },
      "SaveBadRequestBody": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/SchemaErrorBody"
          },
          {
            "$ref": "#/components/schemas/MessageBody"
          }
        ]
      },
...
    }
  }
}